make typedefs compatible with freeimage

+++ dcraw/dcraw.c
@@ -80,8 +80,9 @@ typedef unsigned __int64 UINT64;
 #include <unistd.h>
 #include <utime.h>
 #include <netinet/in.h>
-typedef long long INT64;
-typedef unsigned long long UINT64;
+#include <inttypes.h>
+typedef int64_t INT64;
+typedef uint64_t UINT64;
 #endif
 
 #ifdef NODEPS
